+++ examples/AutoBenchmark/generate_readme.py
@@ -40,7 +40,7 @@ is the overhead caused by the `Coroutine` context switch.
 
 All times in below are in microseconds.
 
-**Version**: AceRoutine v1.2.4
+**Version**: AceRoutine v1.3
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -86,7 +86,30 @@ $ make README.md
 
 ## CPU Time Changes
 
-Version 1.2.3 adds benchmarks for STM32.
+* v1.2.3
+    * Add benchmarks for STM32.
+* v1.3
+    * Replace floating point calculation in AutoBenchmark with fixed point
+      calculations in micros and nanos, with the final printing done in
+      micros to 3 decimal places (without using floating point ops).
+    * Replace looping over a fixed elapsed time, with looping over fixed number
+      of iterations for better accuracy.
+        * If the elapsed time is kept constant, then the number of iteration
+          of `doBaseline()` is different than the number of iterations
+          of doAceRoutine()`. When we subtract, the overhead of the loop
+          (e.g. `millis() - start`) are NOT canceld out correctly.
+        * New numbers
+            * Nano: 5.28 -> 6.500 micros
+            * Micro: 5.31 -> 6.500 micros
+            * SAMD: 2.46 -> 2.333 micros
+            * STM32: 1.76 -> 1.767 micros
+            * ESP8266: 1.67 -> 1.500 micros
+            * ESP32: 0.41 -> 0.400 micros
+            * Teensy 3.2: 1.01 -> 1.000 micros
+    * Replace virtual clock ticking methods (`Coroutine::coroutineMillis()`,
+      `Coroutine::coroutineMicros()`, `Coroutine::coroutineSeconds()`) with
+      static calls to `ClockInterface` template class.
+        * 10-40% performance improvement of `CoroutineScheduler::loop()`.
 
 ## Arduino Nano