+++ examples/MemoryBenchmark/generate_readme.py
@@ -38,7 +38,7 @@ calculated flash size can jump around in unexpected ways.
 
 **NOTE**: This file was auto-generated using `make README.md`. DO NOT EDIT.
 
-**Version**: AceRoutine v1.2
+**Version**: AceRoutine v1.3
 
 **Changes**:
 
@@ -51,6 +51,24 @@ calculated flash size can jump around in unexpected ways.
       linked list
     * `CoroutineScheduler::setupScheduler()` simplified due to immutable linked
       list
+* v1.3
+    * Remove virtual destructor on `Coroutine` class. Reduces flash memory
+      consumption by 500-600 bytes on AVR processors, 350 bytes on SAMD21, and
+      50-150 bytes on other 32-bit processors. The static memory is also reduced
+      by 14 bytes on AVR processors.
+    * Replace clock ticking virtual methods (`coroutineMicros()`,
+      `coroutineMillis()`, and `coroutineSeconds()`) with static functions that
+      delegate to `ClockInterface` which is a template parameter. Saves only
+      0-40 bytes of flash on on AVR processors, but 100-1500 bytes of flash on
+      32-bit processors.
+    * Add benchmark for 'Manual Delay Loop' which uses a simple function
+      to implement the functionality of a `COROUTINE()` that loops every 10
+      milliseconds.
+        * Illustrates clearly that AceRoutine should rarely be used on 8-bit
+          processors because it consumes 450 extra bytes for the 1st coroutine,
+          and another 300 bytes for each additional coroutine.
+        * On 32-bit processors with large amount of flash memory, the flash
+          consumption overhead is not as limiting.
 
 ## How to Generate